Se cambio el  tipo de datos de decimal a entero tabla DatosFisicos

+++ health/models/datos_fisicos_models.py
@@ -5,7 +5,7 @@ from users.models.usuario_models import Usuario
 class DatosFisicos(models.Model):
     usuario = models.ForeignKey(Usuario, on_delete=models.CASCADE)
     peso = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
-    altura = models.DecimalField(max_digits=4, decimal_places=2, null=True, blank=True)
+    altura = models.IntegerField(null=True, blank=True)
     imc = models.DecimalField(max_digits=4, decimal_places=2, null=True, blank=True)
     radio_abdominal = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
     grasa_corporal = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
